Relative path chromedrive | load fallback phrases | improvement file iterations

+++ StoneTest.Crawler.App/StoneTest.Crawler.DataModule/Interfaces/IFileManager.cs
@@ -4,13 +4,13 @@ namespace StoneTest.Crawler.PersistenceModule.Interfaces
 {
     public interface IFileManager
     {
-        public double CurrentFileSizeMB { get; }
+        public decimal CurrentFileSizeMB { get; }
         public string CurrentFileName { get; }
 
         public string FilePath { get; }
         public string FileName { get; }
 
-        public double WriteContent(TextContent content);
+        public decimal WriteContent(TextContent content);
 
         public string ReadContent();
     }
